Define DecoupledSubgraph & Subgraph in same file
Avoids circular imports more elegantly than deferring the import.
